WIP refactoring some methods
